+++ src/streetworks/common/__init__.py
@@ -42,6 +42,7 @@ from .from_nzta import from_nzta
 from .from_ogc_features import from_ogc_features
 from .from_openusrn import from_openusrn
 from .from_oslo import from_oslo
+from .from_osni import from_osni
 from .from_paris import from_paris
 from .from_roma import from_roma
 from .from_sct import from_sct
@@ -150,4 +151,5 @@ __all__ = [
     "from_vienna",
     "from_tfl",
     "from_idee",
+    "from_osni",
 ]